Remove options of using APPENGINE_CONFIG
Users of all types of deployments must use either
1. An actual value
2. "GCLOUD_CONFIG" to delegate to gcloud's configured state.
