+++ src/main/java/com/google/cloud/tools/gradle/appengine/appyaml/AppEngineAppYamlPlugin.java
@@ -108,8 +108,6 @@ public class AppEngineAppYamlPlugin implements Plugin<Project> {
             deploy.setAppEngineDirectory(stageExtension.getAppEngineDirectory());
           }
 
-          deploy.setDeployTargetResolver(new AppYamlDeployTargetResolver(cloudSdkOperations));
-
           DeployAllTask deployAllTask =
               (DeployAllTask)
                   project
